// frontend/Angular/Angular/src/app/clients/users-api.client.ts
export class SignInDto implements ISignInDto {
    email!: string | undefined;
    password!: string | undefined;

    constructor(data?: ISignInDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
            this.password = _data["password"];
        }
    }

    static fromJS(data: any): SignInDto {
        data = typeof data === 'object' ? data : {};
        let result = new SignInDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["password"] = this.password;
        return data;
    }
}

export interface ISignInDto {
    email: string | undefined;
    password: string | undefined;
}

export class SignInResultDto implements ISignInResultDto {
    succeeded!: boolean;
    errorType!: number | undefined;
    token!: string | undefined;

    constructor(data?: ISignInResultDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.succeeded = _data["succeeded"];
            this.errorType = _data["errorType"];
            this.token = _data["token"];
        }
    }

    static fromJS(data: any): SignInResultDto {
        data = typeof data === 'object' ? data : {};
        let result = new SignInResultDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["succeeded"] = this.succeeded;
        data["errorType"] = this.errorType;
        data["token"] = this.token;
        return data;
    }
}

export interface ISignInResultDto {
    succeeded: boolean;
    errorType: number | undefined;
    token: string | undefined;
}

export class SignUpDto implements ISignUpDto {
    name!: string | undefined;
    age!: number;
    email!: string | undefined;
    password!: string | undefined;

    constructor(data?: ISignUpDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.age = _data["age"];
            this.email = _data["email"];
            this.password = _data["password"];
        }
    }

    static fromJS(data: any): SignUpDto {
        data = typeof data === 'object' ? data : {};
        let result = new SignUpDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["age"] = this.age;
        data["email"] = this.email;
        data["password"] = this.password;
        return data;
    }
}

export interface ISignUpDto {
    name: string | undefined;
    age: number;
    email: string | undefined;
    password: string | undefined;
}